Tests refactor (#208)
* saving current state of the refactoring

* Making Spark Session a class-level fixture
Suppressing unhelpful warnings

* Pandas baseline not ready yet.
