+++ python/tests/tsdf_tests.py
@@ -1,111 +1,11 @@
-import logging
 import unittest
-import re
 
 import pyspark.sql.functions as F
-from pyspark.sql import SparkSession
 from pyspark.sql.types import *
 
 from tempo.tsdf import TSDF
-from tempo.utils import *
 
-class SparkTest(unittest.TestCase):
-    ##
-    ## Fixtures
-    ##
-    def setUp(self):
-        self.spark = (SparkSession.builder.appName("myapp") \
-                      .config("spark.jars.packages", "io.delta:delta-core_2.12:1.1.0") \
-                      .config("spark.sql.extensions", "io.delta.sql.DeltaSparkSessionExtension") \
-                      .config("spark.sql.catalog.spark_catalog", "org.apache.spark.sql.delta.catalog.DeltaCatalog") \
-                      .config("spark.driver.extraJavaOptions", "-Dio.netty.tryReflectionSetAccessible=true") \
-                      .config("spark.executor.extraJavaOptions", "-Dio.netty.tryReflectionSetAccessible=true") \
-                      .master("local") \
-                      .getOrCreate())
-        self.spark.conf.set("spark.sql.shuffle.partitions", 1)
-
-    def tearDown(self) -> None:
-        self.spark.stop()
-
-    ##
-    ## Utility Functions
-    ##
-
-    def buildTestDF(self, schema, data, ts_cols=["event_ts"]):
-        """
-        Constructs a Spark Dataframe from the given components
-        :param schema: the schema to use for the Dataframe
-        :param data: values to use for the Dataframe
-        :param ts_cols: list of column names to be converted to Timestamp values
-        :return: a Spark Dataframe, constructed from the given schema and values
-        """
-        # build dataframe
-        df = self.spark.createDataFrame(data, schema)
-
-        #check if ts_col follows standard timestamp format, then check if timestamp has micro/nanoseconds
-        for tsc in ts_cols:
-            ts_value = str(df.select(ts_cols).limit(1).collect()[0][0])
-            ts_pattern = "^\d{4}-\d{2}-\d{2}| \d{2}:\d{2}:\d{2}\.\d*$"
-            decimal_pattern = "[.]\d+"
-            if re.match(ts_pattern, str(ts_value)) is not None:
-                if re.search(decimal_pattern, ts_value) is None or len(re.search(decimal_pattern, ts_value)[0]) <= 4:
-                    df = df.withColumn(tsc, F.to_timestamp(F.col(tsc)))
-        return df
-
-    ##
-    ## DataFrame Assert Functions
-    ##
-
-    def assertFieldsEqual(self, fieldA, fieldB):
-        """
-        Test that two fields are equivalent
-        """
-        self.assertEqual(fieldA.name.lower(), fieldB.name.lower())
-        self.assertEqual(fieldA.dataType, fieldB.dataType)
-        # self.assertEqual(fieldA.nullable, fieldB.nullable)
-
-    def assertSchemaContainsField(self, schema, field):
-        """
-        Test that the given schema contains the given field
-        """
-        # the schema must contain a field with the right name
-        lc_fieldNames = [fc.lower() for fc in schema.fieldNames()]
-        self.assertTrue(field.name.lower() in lc_fieldNames)
-        # the attributes of the fields must be equal
-        self.assertFieldsEqual(field, schema[field.name])
-
-    def assertSchemasEqual(self, schemaA, schemaB):
-        """
-        Test that the two given schemas are equivalent (column ordering ignored)
-        """
-        # both schemas must have the same length
-        self.assertEqual(len(schemaA.fields), len(schemaB.fields))
-        # schemaA must contain every field in schemaB
-        for field in schemaB.fields:
-            self.assertSchemaContainsField(schemaA, field)
-
-    def assertHasSchema(self, df, expectedSchema):
-        """
-        Test that the given Dataframe conforms to the expected schema
-        """
-        self.assertSchemasEqual(df.schema, expectedSchema)
-
-    def assertDataFramesEqual(self, dfA, dfB):
-        """
-        Test that the two given Dataframes are equivalent.
-        That is, they have equivalent schemas, and both contain the same values
-        """
-        # must have the same schemas
-        self.assertSchemasEqual(dfA.schema, dfB.schema)
-        # enforce a common column ordering
-        colOrder = sorted(dfA.columns)
-        sortedA = dfA.select(colOrder)
-        sortedB = dfB.select(colOrder)
-        # must have identical data
-        # that is all rows in A must be in B, and vice-versa
-
-        self.assertEqual(sortedA.subtract(sortedB).count(), 0)
-        self.assertEqual(sortedB.subtract(sortedA).count(), 0)
+from tests.base import SparkTest
 
 
 class BasicTests(SparkTest):
@@ -162,292 +62,6 @@ class BasicTests(SparkTest):
                    0] == '2020-08-01 00:00:10'
         assert res.filter(F.col("max_ts") != " ").select(F.col('max_ts').cast("string")).collect()[0][
                    0] == '2020-09-01 00:19:12'
-
-
-class AsOfJoinTest(SparkTest):
-
-    def test_asof_join(self):
-        """AS-OF Join with out a time-partition test"""
-        leftSchema = StructType([StructField("symbol", StringType()),
-                                 StructField("event_ts", StringType()),
-                                 StructField("trade_pr", FloatType())])
-
-        rightSchema = StructType([StructField("symbol", StringType()),
-                                  StructField("event_ts", StringType()),
-                                  StructField("bid_pr", FloatType()),
-                                  StructField("ask_pr", FloatType())])
-
-        expectedSchema = StructType([StructField("symbol", StringType()),
-                                     StructField("left_event_ts", StringType()),
-                                     StructField("left_trade_pr", FloatType()),
-                                     StructField("right_event_ts", StringType()),
-                                     StructField("right_bid_pr", FloatType()),
-                                     StructField("right_ask_pr", FloatType())])
-        expectedSchemaNoRightPrefix = StructType([StructField("symbol", StringType()),
-                                     StructField("left_event_ts", StringType()),
-                                     StructField("left_trade_pr", FloatType()),
-                                     StructField("event_ts", StringType()),
-                                     StructField("bid_pr", FloatType()),
-                                     StructField("ask_pr", FloatType())])
-
-        left_data = [["S1", "2020-08-01 00:00:10", 349.21],
-                     ["S1", "2020-08-01 00:01:12", 351.32],
-                     ["S1", "2020-09-01 00:02:10", 361.1],
-                     ["S1", "2020-09-01 00:19:12", 362.1]]
-
-        right_data = [["S1", "2020-08-01 00:00:01", 345.11, 351.12],
-                      ["S1", "2020-08-01 00:01:05", 348.10, 353.13],
-                      ["S1", "2020-09-01 00:02:01", 358.93, 365.12],
-                      ["S1", "2020-09-01 00:15:01", 359.21, 365.31]]
-
-        expected_data = [
-            ["S1", "2020-08-01 00:00:10", 349.21, "2020-08-01 00:00:01", 345.11, 351.12],
-            ["S1", "2020-08-01 00:01:12", 351.32, "2020-08-01 00:01:05", 348.10, 353.13],
-            ["S1", "2020-09-01 00:02:10", 361.1, "2020-09-01 00:02:01", 358.93, 365.12],
-            ["S1", "2020-09-01 00:19:12", 362.1, "2020-09-01 00:15:01", 359.21, 365.31]]
-
-        # Construct dataframes
-        dfLeft = self.buildTestDF(leftSchema, left_data)
-        dfRight = self.buildTestDF(rightSchema, right_data)
-        dfExpected = self.buildTestDF(expectedSchema, expected_data, ["left_event_ts", "right_event_ts"])
-
-        # perform the join
-        tsdf_left = TSDF(dfLeft, ts_col="event_ts", partition_cols=["symbol"])
-        tsdf_right = TSDF(dfRight, ts_col="event_ts", partition_cols=["symbol"])
-
-        joined_df = tsdf_left.asofJoin(tsdf_right, left_prefix="left", right_prefix="right").df
-        non_prefix_joined_df = tsdf_left.asofJoin(tsdf_right, left_prefix="left", right_prefix = '').df
-
-        # joined dataframe should equal the expected dataframe
-        self.assertDataFramesEqual(joined_df, dfExpected)
-
-        noRightPrefixdfExpected = self.buildTestDF(expectedSchemaNoRightPrefix, expected_data, ["left_event_ts", "event_ts"])
-
-        self.assertDataFramesEqual(non_prefix_joined_df, noRightPrefixdfExpected)
-
-        spark_sql_joined_df = tsdf_left.asofJoin(tsdf_right, left_prefix="left", right_prefix="right").df
-        self.assertDataFramesEqual(spark_sql_joined_df, dfExpected)
-
-    def test_asof_join_skip_nulls_disabled(self):
-        """AS-OF Join with skip nulls disabled"""
-        leftSchema = StructType([StructField("symbol", StringType()),
-                                 StructField("event_ts", StringType()),
-                                 StructField("trade_pr", FloatType())])
-
-        rightSchema = StructType([StructField("symbol", StringType()),
-                                  StructField("event_ts", StringType()),
-                                  StructField("bid_pr", FloatType()),
-                                  StructField("ask_pr", FloatType())])
-
-        expectedSchema = StructType([StructField("symbol", StringType()),
-                                     StructField("left_event_ts", StringType()),
-                                     StructField("left_trade_pr", FloatType()),
-                                     StructField("right_event_ts", StringType()),
-                                     StructField("right_bid_pr", FloatType()),
-                                     StructField("right_ask_pr", FloatType())])
-
-        left_data = [["S1", "2020-08-01 00:00:10", 349.21],
-                     ["S1", "2020-08-01 00:01:12", 351.32],
-                     ["S1", "2020-09-01 00:02:10", 361.1],
-                     ["S1", "2020-09-01 00:19:12", 362.1]]
-
-        right_data = [["S1", "2020-08-01 00:00:01", 345.11, 351.12],
-                      ["S1", "2020-08-01 00:01:05", None, 353.13],
-                      ["S1", "2020-09-01 00:02:01", None, None],
-                      ["S1", "2020-09-01 00:15:01", 359.21, 365.31]]
-
-        expected_data_skip_nulls = [
-            ["S1", "2020-08-01 00:00:10", 349.21, "2020-08-01 00:00:01", 345.11, 351.12],
-            ["S1", "2020-08-01 00:01:12", 351.32, "2020-08-01 00:01:05", 345.11, 353.13],
-            ["S1", "2020-09-01 00:02:10", 361.1, "2020-09-01 00:02:01", 345.11, 353.13],
-            ["S1", "2020-09-01 00:19:12", 362.1, "2020-09-01 00:15:01", 359.21, 365.31]]
-
-        expected_data_skip_nulls_disabled = [
-            ["S1", "2020-08-01 00:00:10", 349.21, "2020-08-01 00:00:01", 345.11, 351.12],
-            ["S1", "2020-08-01 00:01:12", 351.32, "2020-08-01 00:01:05", None, 353.13],
-            ["S1", "2020-09-01 00:02:10", 361.1, "2020-09-01 00:02:01", None, None],
-            ["S1", "2020-09-01 00:19:12", 362.1, "2020-09-01 00:15:01", 359.21, 365.31]]
-
-        # Construct dataframes
-        dfLeft = self.buildTestDF(leftSchema, left_data)
-        dfRight = self.buildTestDF(rightSchema, right_data)
-        dfExpectedSkipNulls = self.buildTestDF(
-            expectedSchema, expected_data_skip_nulls, ["left_event_ts", "right_event_ts"]
-        )
-        dfExpectedSkipNullsDisabled = self.buildTestDF(
-            expectedSchema, expected_data_skip_nulls_disabled, ["left_event_ts", "right_event_ts"]
-        )
-
-        tsdf_left = TSDF(dfLeft, ts_col="event_ts", partition_cols=["symbol"])
-        tsdf_right = TSDF(dfRight, ts_col="event_ts", partition_cols=["symbol"])
-
-        # perform the join with skip nulls enabled (default)
-        joined_df = tsdf_left.asofJoin(tsdf_right, left_prefix="left", right_prefix="right").df
-
-        # joined dataframe should equal the expected dataframe with nulls skipped
-        self.assertDataFramesEqual(joined_df, dfExpectedSkipNulls)
-
-        # perform the join with skip nulls disabled
-        joined_df = tsdf_left.asofJoin(tsdf_right, left_prefix="left", right_prefix="right", skipNulls=False).df
-
-        # joined dataframe should equal the expected dataframe without nulls skipped
-        self.assertDataFramesEqual(joined_df, dfExpectedSkipNullsDisabled)
-
-    def test_sequence_number_sort(self):
-        """Skew AS-OF Join with Partition Window Test"""
-        leftSchema = StructType([StructField("symbol", StringType()),
-                                 StructField("event_ts", StringType()),
-                                 StructField("trade_pr", FloatType()),
-                                 StructField("trade_id", IntegerType())])
-
-        rightSchema = StructType([StructField("symbol", StringType()),
-                                  StructField("event_ts", StringType()),
-                                  StructField("bid_pr", FloatType()),
-                                  StructField("ask_pr", FloatType()),
-                                  StructField("seq_nb", LongType())])
-
-        expectedSchema = StructType([StructField("symbol", StringType()),
-                                     StructField("event_ts", StringType()),
-                                     StructField("trade_pr", FloatType()),
-                                     StructField("trade_id", IntegerType()),
-                                     StructField("right_event_ts", StringType()),
-                                     StructField("right_bid_pr", FloatType()),
-                                     StructField("right_ask_pr", FloatType()),
-                                     StructField("right_seq_nb", LongType())])
-
-        left_data = [["S1", "2020-08-01 00:00:10", 349.21, 1],
-                     ["S1", "2020-08-01 00:00:10", 350.21, 5],
-                     ["S1", "2020-08-01 00:01:12", 351.32, 2],
-                     ["S1", "2020-09-01 00:02:10", 361.1, 3],
-                     ["S1", "2020-09-01 00:19:12", 362.1, 4]]
-
-        right_data = [["S1", "2020-08-01 00:00:01", 345.11, 351.12, 1],
-                      ["S1", "2020-08-01 00:00:10", 19.11, 20.12, 1],
-                      ["S1", "2020-08-01 00:01:05", 348.10, 1000.13, 3],
-                      ["S1", "2020-08-01 00:01:05", 348.10, 100.13, 2],
-                      ["S1", "2020-09-01 00:02:01", 358.93, 365.12, 4],
-                      ["S1", "2020-09-01 00:15:01", 359.21, 365.31, 5]]
-
-        expected_data = [
-            ["S1", "2020-08-01 00:00:10", 349.21, 1, "2020-08-01 00:00:10", 19.11, 20.12, 1],
-            ["S1", "2020-08-01 00:00:10", 350.21, 5, "2020-08-01 00:00:10", 19.11, 20.12, 1],
-            ["S1", "2020-08-01 00:01:12", 351.32, 2, "2020-08-01 00:01:05", 348.10, 1000.13, 3],
-            ["S1", "2020-09-01 00:02:10", 361.1, 3, "2020-09-01 00:02:01", 358.93, 365.12, 4],
-            ["S1", "2020-09-01 00:19:12", 362.1, 4, "2020-09-01 00:15:01", 359.21, 365.31, 5]]
-
-        # construct dataframes
-        dfLeft = self.buildTestDF(leftSchema, left_data)
-        dfRight = self.buildTestDF(rightSchema, right_data)
-        dfExpected = self.buildTestDF(expectedSchema, expected_data, ["right_event_ts", "event_ts"])
-
-        # perform the join
-        tsdf_left = TSDF(dfLeft, partition_cols=["symbol"])
-        tsdf_right = TSDF(dfRight, partition_cols=["symbol"], sequence_col="seq_nb")
-        joined_df = tsdf_left.asofJoin(tsdf_right, right_prefix='right').df
-
-        # joined dataframe should equal the expected dataframe
-        self.assertDataFramesEqual(joined_df, dfExpected)
-
-    def test_partitioned_asof_join(self):
-        """AS-OF Join with a time-partition"""
-        leftSchema = StructType([StructField("symbol", StringType()),
-                                 StructField("event_ts", StringType()),
-                                 StructField("trade_pr", FloatType())])
-
-        rightSchema = StructType([StructField("symbol", StringType()),
-                                  StructField("event_ts", StringType()),
-                                  StructField("bid_pr", FloatType()),
-                                  StructField("ask_pr", FloatType())])
-
-        expectedSchema = StructType([StructField("symbol", StringType()),
-                                     StructField("left_event_ts", StringType()),
-                                     StructField("left_trade_pr", FloatType()),
-                                     StructField("right_event_ts", StringType()),
-                                     StructField("right_bid_pr", FloatType()),
-                                     StructField("right_ask_pr", FloatType())])
-
-        left_data = [["S1", "2020-08-01 00:00:02", 349.21],
-                     ["S1", "2020-08-01 00:00:08", 351.32],
-                     ["S1", "2020-08-01 00:00:11", 361.12],
-                     ["S1", "2020-08-01 00:00:18", 364.31],
-                     ["S1", "2020-08-01 00:00:19", 362.94],
-                     ["S1", "2020-08-01 00:00:21", 364.27],
-                     ["S1", "2020-08-01 00:00:23", 367.36]]
-
-        right_data = [["S1", "2020-08-01 00:00:01", 345.11, 351.12],
-                      ["S1", "2020-08-01 00:00:09", 348.10, 353.13],
-                      ["S1", "2020-08-01 00:00:12", 358.93, 365.12],
-                      ["S1", "2020-08-01 00:00:19", 359.21, 365.31]]
-
-        expected_data = [
-            ["S1", "2020-08-01 00:00:02", 349.21, "2020-08-01 00:00:01", 345.11, 351.12],
-            ["S1", "2020-08-01 00:00:08", 351.32, "2020-08-01 00:00:01", 345.11, 351.12],
-            ["S1", "2020-08-01 00:00:11", 361.12, "2020-08-01 00:00:09", 348.10, 353.13],
-            ["S1", "2020-08-01 00:00:18", 364.31, "2020-08-01 00:00:12", 358.93, 365.12],
-            ["S1", "2020-08-01 00:00:19", 362.94, "2020-08-01 00:00:19", 359.21, 365.31],
-            ["S1", "2020-08-01 00:00:21", 364.27, "2020-08-01 00:00:19", 359.21, 365.31],
-            ["S1", "2020-08-01 00:00:23", 367.36, "2020-08-01 00:00:19", 359.21, 365.31]]
-
-        # Construct dataframes
-        dfLeft = self.buildTestDF(leftSchema, left_data)
-        dfRight = self.buildTestDF(rightSchema, right_data)
-        dfExpected = self.buildTestDF(expectedSchema, expected_data, ["left_event_ts", "right_event_ts"])
-
-        tsdf_left = TSDF(dfLeft, ts_col="event_ts", partition_cols=["symbol"])
-        tsdf_right = TSDF(dfRight, ts_col="event_ts", partition_cols=["symbol"])
-
-        joined_df = tsdf_left.asofJoin(tsdf_right, left_prefix="left", right_prefix="right",
-                                       tsPartitionVal=10, fraction=0.1).df
-
-        self.assertDataFramesEqual(joined_df, dfExpected)
-
-    def test_asof_join_nanos(self):
-        """As of join with nanosecond timestamps"""
-        leftSchema = StructType([StructField("symbol", StringType()),
-                                 StructField("event_ts", StringType()),
-                                 StructField("trade_pr", FloatType())])
-
-        rightSchema = StructType([StructField("symbol", StringType()),
-                                  StructField("event_ts", StringType()),
-                                  StructField("bid_pr", FloatType()),
-                                  StructField("ask_pr", FloatType())])
-
-        expectedSchema = StructType([StructField("symbol", StringType()),
-                                     StructField("left_event_ts", StringType()),
-                                     StructField("left_trade_pr", FloatType()),
-                                     StructField("right_event_ts", StringType()),
-                                     StructField("right_ask_pr", FloatType()),
-                                     StructField("right_bid_pr", FloatType())]
-                                     )
-
-        left_data = [["S1", "2022-01-01 09:59:59.123456789", 349.21],
-                     ["S1", "2022-01-01 10:00:00.123456788", 351.32],
-                     ["S1", "2022-01-01 10:00:00.123456789", 361.12],
-                     ["S1", "2022-01-01 10:00:01.123456789", 364.31], ]
-
-        right_data = [["S1", "2022-01-01 10:00:00.1234567", 345.11, 351.12],
-                      ["S1", "2022-01-01 10:00:00.12345671", 348.10, 353.13],
-                      ["S1", "2022-01-01 10:00:00.12345675", 358.93, 365.12],
-                      ["S1", "2022-01-01 10:00:00.12345677", 358.91, 365.33],
-                      ["S1", "2022-01-01 10:00:01.10000001", 359.21, 365.31]]
-
-        expected_data = [
-            ["S1", "2022-01-01 09:59:59.123456789", 349.21, None, None, None],
-            ["S1", "2022-01-01 10:00:00.123456788", 351.32, "2022-01-01 10:00:00.12345677", 365.33, 358.91],
-            ["S1", "2022-01-01 10:00:00.123456789", 361.12, "2022-01-01 10:00:00.12345677", 365.33, 358.91],
-            ["S1", "2022-01-01 10:00:01.123456789", 364.31, "2022-01-01 10:00:01.10000001", 365.31, 359.21]]
-
-        dfLeft = self.buildTestDF(leftSchema, left_data)
-        dfRight = self.buildTestDF(rightSchema, right_data)
-        dfExpected = self.buildTestDF(expectedSchema, expected_data, ts_cols=["left_event_ts"])
-
-        tsdf_left = TSDF(dfLeft, ts_col="event_ts", partition_cols=["symbol"])
-        tsdf_right = TSDF(dfRight, ts_col="event_ts", partition_cols=["symbol"])
-
-        joined_df = tsdf_left.asofJoin(tsdf_right, left_prefix="left", right_prefix="right").df
-
-        self.assertDataFramesEqual(joined_df, dfExpected)
-
 
 class FourierTransformTest(SparkTest):
 
@@ -832,38 +446,6 @@ class ResampleTest(SparkTest):
         self.assertDataFramesEqual(bars, barsExpected)
 
 
-class DeltaWriteTest(SparkTest):
-
-    def test_write_to_delta(self):
-        """Test table write to delta format"""
-        schema = StructType([StructField("symbol", StringType()),
-                             StructField("date", StringType()),
-                             StructField("event_ts", StringType()),
-                             StructField("trade_pr", FloatType()),
-                             StructField("trade_pr_2", FloatType())])
-
-        data = [["S1", "SAME_DT", "2020-08-01 00:00:10", 349.21, 10.0],
-                ["S1", "SAME_DT", "2020-08-01 00:00:11", 340.21, 9.0],
-                ["S1", "SAME_DT", "2020-08-01 00:01:12", 353.32, 8.0],
-                ["S1", "SAME_DT", "2020-08-01 00:01:13", 351.32, 7.0],
-                ["S1", "SAME_DT", "2020-08-01 00:01:14", 350.32, 6.0],
-                ["S1", "SAME_DT", "2020-09-01 00:01:12", 361.1, 5.0],
-                ["S1", "SAME_DT", "2020-09-01 00:19:12", 362.1, 4.0]]
-
-        # construct dataframe
-        df = self.buildTestDF(schema, data)
-
-        # convert to TSDF
-        tsdf_left = TSDF(df, partition_cols=["symbol"], ts_col="event_ts")
-
-        #test write to delta
-        tsdf_left.write(self.spark, "my_table")
-        logging.info('delta table count ' + str(self.spark.table("my_table").count()))
-
-        # should be equal to the expected dataframe
-        assert self.spark.table("my_table").count() == 7
-
-
-## MAIN
+# MAIN
 if __name__ == '__main__':
     unittest.main()
